# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AugmentedQuill",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_name}/annotations": {
      "get": {
        "tags": ["Annotations"],
        "summary": "Api List Annotations",
        "description": "List all annotations for the project.",
        "operationId": "api_list_annotations_api_v1_projects__project_name__annotations_get",
        "parameters": [
          {
            "name": "project_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Directory name of the project",
              "title": "Project Name"
            },
            "description": "Directory name of the project"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnnotationResponse"
                  },
                  "title": "Response Api List Annotations Api V1 Projects  Project Name  Annotations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Annotations"],
        "summary": "Api Create Annotation",
        "description": "Create a new inline annotation.",
        "operationId": "api_create_annotation_api_v1_projects__project_name__annotations_post",
        "parameters": [
          {
            "name": "project_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Directory name of the project",
              "title": "Project Name"
            },
            "description": "Directory name of the project"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAnnotationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnotationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_name}/annotations/scope": {
      "get": {
        "tags": ["Annotations"],
        "summary": "Api List Annotations For Scope",
        "description": "List annotations for a specific prose scope.",
        "operationId": "api_list_annotations_for_scope_api_v1_projects__project_name__annotations_scope_get",
        "parameters": [
          {
            "name": "project_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Directory name of the project",
              "title": "Project Name"
            },
            "description": "Directory name of the project"
          },
          {
            "name": "scope_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "story",
              "title": "Scope Type"
            }
          },
          {
            "name": "chapter_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Chapter Id"
            }
          },
          {
            "name": "book_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Book Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnnotationResponse"
                  },
                  "title": "Response Api List Annotations For Scope Api V1 Projects  Project Name  Annotations Scope Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_name}/annotations/{annotation_id}": {
      "get": {
        "tags": ["Annotations"],
        "summary": "Api Get Annotation",
        "description": "Fetch a single annotation by ID.",
        "operationId": "api_get_annotation_api_v1_projects__project_name__annotations__annotation_id__get",
        "parameters": [
          {
            "name": "annotation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Annotation Id"
            }
          },
          {
            "name": "project_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Directory name of the project",
              "title": "Project Name"
            },
            "description": "Directory name of the project"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnotationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Annotations"],
        "summary": "Api Update Annotation",
        "description": "Update an annotation's comment.",
        "operationId": "api_update_annotation_api_v1_projects__project_name__annotations__annotation_id__put",
        "parameters": [
          {
            "name": "annotation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Annotation Id"
            }
          },
          {
            "name": "project_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Directory name of the project",
              "title": "Project Name"
            },
            "description": "Directory name of the project"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAnnotationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnotationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Annotations"],
        "summary": "Api Delete Annotation",
        "description": "Delete an annotation and remove its inline markers.",
        "operationId": "api_delete_annotation_api_v1_projects__project_name__annotations__annotation_id__delete",
        "parameters": [
          {
            "name": "annotation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Annotation Id"
            }
          },
          {
            "name": "project_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Directory name of the project",
              "title": "Project Name"
            },
            "description": "Directory name of the project"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnnotationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "comment": {
            "type": "string",
            "title": "Comment"
          },
          "scope_type": {
            "type": "string",
            "title": "Scope Type"
          },
          "chapter_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter Id"
          },
          "book_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Book Id"
          },
          "start_offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Offset"
          },
          "end_offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Offset"
          }
        },
        "type": "object",
        "required": ["id", "comment", "scope_type"],
        "title": "AnnotationResponse",
        "description": "A single annotation with runtime offsets attached."
      },
...
      "CreateAnnotationRequest": {
        "properties": {
          "scope_type": {
            "type": "string",
            "title": "Scope Type",
            "description": "'story', 'chapter', or 'unlinked'"
          },
          "chapter_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter Id",
            "description": "Chapter ID for chapter scope"
          },
          "book_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Book Id",
            "description": "Book ID for nested chapter scope"
          },
          "start_offset": {
            "type": "integer",
            "title": "Start Offset",
            "description": "Raw character offset (markers included)"
          },
          "end_offset": {
            "type": "integer",
            "title": "End Offset",
            "description": "Raw character offset (markers included)"
          },
          "comment": {
            "type": "string",
            "title": "Comment",
            "description": "Human-readable annotation comment"
          }
        },
        "type": "object",
        "required": ["scope_type", "start_offset", "end_offset", "comment"],
        "title": "CreateAnnotationRequest",
        "description": "Payload for creating a new annotation."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UpdateAnnotationRequest": {
        "properties": {
          "comment": {
            "type": "string",
            "title": "Comment"
          }
        },
        "type": "object",
        "required": ["comment"],
        "title": "UpdateAnnotationRequest",
        "description": "Payload for updating an annotation's comment."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    }
  }
}